+++ tablet_qt/menu/researchmenu.cpp
@@ -30,6 +30,7 @@
 #include "tasks/chit.h"
 #include "tasks/diagnosisicd9cm.h"
 #include "tasks/ided3d.h"
+#include "tasks/isaaq.h"
 #include "tasks/kirby.h"
 #include "tasks/mfi20.h"
 #include "tasks/qolbasic.h"
@@ -75,7 +76,7 @@ void ResearchMenu::makeItems()
         ),
         MAKE_TASK_MENU_ITEM(CardinalExpDetThreshold::CARDINALEXPDETTHRESHOLD_TABLENAME, m_app),
         MAKE_TASK_MENU_ITEM(CardinalExpectationDetection::CARDINALEXPDET_TABLENAME, m_app),
-        // todo:: ISAAQ
+        MAKE_TASK_MENU_ITEM(Isaaq::ISAAQ_TABLENAME, m_app),
     };
 }